# Add mkl implementation for exponential on CPU