--- a/test/distributions/test_distributions.py
+++ b/test/distributions/test_distributions.py
@@ -2411,6 +2411,18 @@ class TestDistributions(DistributionsTestCase):
         self._check_log_prob(Exponential(rate), ref_log_prob)
         self._check_forward_ad(lambda x: x.exponential_())
 
+        def mean_var(lambd, sample):
+            sample.exponential_(lambd)
+            mean = sample.mean()
+            var = sample.var()
+            self.assertEqual((1. / lambd), mean, atol=1e-2, rtol=1e-2)
+            self.assertEqual((1. / lambd) ** 2, var, atol=2e-2, rtol=1e-2)
+
+        for dtype in [torch.float, torch.double]:
+            for lambd in [0.2, 0.5, 1., 1.5, 2., 5.]:
+                sample_len = 50000
+                mean_var(lambd, torch.rand(sample_len, dtype=dtype))
+
     @unittest.skipIf(not TEST_NUMPY, "NumPy not found")
     def test_exponential_sample(self):
         set_rng_seed(1)  # see Note [Randomized statistical tests]