--- a/test/distributions/test_distributions.py
+++ b/test/distributions/test_distributions.py
@@ -5075,10 +5075,7 @@ class TestJit(DistributionsTestCase):
             xfail = [
                 Cauchy,  # aten::cauchy(Double(2,1), float, float, Generator)
                 HalfCauchy,  # aten::cauchy(Double(2, 1), float, float, Generator)
-                VonMises,  # Variance is not Euclidean
-                Exponential,  # mkl implementation path on intel cpu will produce diffrent results from jit.trace
-                Pareto,  # base_distribution is Exponential
-                Weibull  # base_distribution is Exponential
+                VonMises  # Variance is not Euclidean
             ]
             if Dist in xfail:
                 continue
@@ -5109,9 +5106,6 @@ class TestJit(DistributionsTestCase):
             xfail = [
                 Cauchy,  # aten::cauchy(Double(2,1), float, float, Generator)
                 HalfCauchy,  # aten::cauchy(Double(2, 1), float, float, Generator)
-                Exponential,  # mkl implementation path on intel cpu will produce diffrent results from jit.trace
-                Pareto,  # base_distribution is Exponential
-                Weibull  # base_distribution is Exponential
             ]
             if Dist in xfail:
                 continue