@malfet 2022-09-08
Also, is there any documentation/results on the spatial autocorrelation / statistical properties of new RNG algorithm vs the old one?