@CaoE 2022-09-15
> Also, is there any documentation/results on the spatial autocorrelation / statistical properties of new RNG algorithm vs the old one?

Hi @malfet  sorry for the late reply ! I was occupied by other tasks. I will post the statistical properties of the new RNG algorithm vs the old one. Also, I didn't find relevant statistical tests on PyTorch for these distribution operators. Do we have plan to add statistical tests for these ops.