@malfet 2022-09-15
> Heads up, @CaoE, unfortunately I have to revert this PR, as it caused breakage of internal Meta tests. I'll provide more details soon.

To followup on the promise: after this change got landed, TestMDETRLosses::test_box_losses from https://github.com/facebookresearch/multimodal started to fail, see https://github.com/facebookresearch/multimodal/blob/afe52dec1987e5a3764539606a55f430994ebbc8/test/modules/losses/test_mdetr_losses.py#L114

Though by looking at the code it feels like it could have been a false positive (I.e. quick look at the code did not shown me where seed is fixed or where tolerance is set correctly.