@CaoE 2022-10-11
Hi @malfet  Tests for mean and var for exponential are added. 

> TestMDETRLosses::test_box_losses from https://github.com/facebookresearch/multimodal started to fail, see https://github.com/facebookresearch/multimodal/blob/afe52dec1987e5a3764539606a55f430994ebbc8/test/modules/losses/test_mdetr_losses.py#L114

https://github.com/facebookresearch/multimodal/blob/afe52dec1987e5a3764539606a55f430994ebbc8/test/modules/losses/test_mdetr_losses.py#L64.  `indices` for  `box_losses` is generated by multinomial which will call exponential, even the new mkl implementation fits the exponential distribution but it is not the same as the old one. Therefore,  fixed `expected_l1_loss = torch.tensor(0.7721)` and `expected_giou_loss = torch.tensor(1.1768)` is not applicable for new exponential.

Can we port the `test_box_losses` to pytorch and add more flexible `expected_l1_loss` and `expected_giou_loss` both in https://github.com/facebookresearch/multimodal/blob/afe52dec1987e5a3764539606a55f430994ebbc8/test/modules/losses/test_mdetr_losses.py and test_distribution.py?