@CaoE 2022-10-27
MKL implementation will also get improvements on AMD machine. I removed the vender check.
![Screenshot (369)](https://user-images.githubusercontent.com/23565213/198245940-346d40b0-87a9-4ff8-89ae-6fff4fd769d4.png)