@jgong5 2022-10-27
@malfet We removed vendor checks in this PR for both bernoulli and exponential distribution kernels since MKL should bring perf benefit for x86 arch in general. Could you please review? Also, please advise AMD point of contact to evaluate the perf impact on AMD CPUs. Thanks.