@CaoE 2022-10-31
@izaitsevfb 
https://github.com/facebookresearch/multimodal/blob/afe52dec1987e5a3764539606a55f430994ebbc8/test/modules/losses/test_mdetr_losses.py#L115. 
May I know why fixed numbers are used here since `test_exponential_sample` in test_distribution.py already uses `scipy.stats` to test sampling results for exponential. 
The fixed `expected_l1_loss = torch.tensor(0.7721)` and `expected_giou_loss = torch.tensor(1.1768)` is not applicable for new exponential.