@CaoE 2022-11-03
@malfet Can we modify the expected loss in https://github.com/facebookresearch/multimodal/blob/afe52dec1987e5a3764539606a55f430994ebbc8/test/modules/losses/test_mdetr_losses.py#L115  to accommodate the new exponential ?