@CaoE 2022-11-23
@malfet @izaitsevfb As suggested by @ebsmothers,  we change the expected value to match the exact value from the new exponential implementationhttps://github.com/facebookresearch/multimodal/pull/364. Could you please check if this PR will cause internal breakage ?