@izaitsevfb 2022-11-29
> @malfet @izaitsevfb As suggested by @ebsmothers, we change the expected value to match the exact value from the new exponential implementationhttps://github.com/[facebookresearch/multimodal/pull/364](https://github.com/facebookresearch/multimodal/pull/364). Could you please check if this PR will cause internal breakage ?

I've manually imported this PR to run the internal tests. For Meta employees, see [D41587318](https://www.internalfb.com/diff/D41587318)