@izaitsevfb 2022-11-30
@CaoE, I've confirmed with the test owners, it's all good. The tests didn't expect the RNG change, but they can be forward fixed. 

For the diff train oncall: please see the details in [D41587318](https://www.internalfb.com/diff/D41587318).