@CaoE 2022-12-05
> @CaoE, I've confirmed with the test owners, it's all good. The tests didn't expect the RNG change, but they can be forward fixed.

Thank you @izaitsevfb. May I know if these tests will be fixed for the new RNG in the near future?