@izaitsevfb 2022-12-06
> > @CaoE, I've confirmed with the test owners, it's all good. The tests didn't expect the RNG change, but they can be forward fixed.
> 
> Thank you @izaitsevfb. May I know if these tests will be fixed for the new RNG in the near future?

I think there are plans to rewrite the tests to not rely on RNG specifics, but I don't have the specifics. Please don't worry about that, as I've confirmed that the internal test failures are expected and can be forward-fixed.