@izaitsevfb 2022-12-13
> > I think there are plans to rewrite the tests to not rely on RNG specifics, but I don't have the specifics. Please don't worry about that, as I've confirmed that the internal test failures are expected and can be forward-fixed.
> 
> @izaitsevfb That's great ! Can we merge this PR first and then fix the tests?

Yep, that's what I meant, sorry for not being clear.